Uralintern/requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-cleanup==5.2.0
django-import-export==2.6.1
django-querycount==0.7.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
fonttools==4.28.2
fpdf==1.7.2
kiwisolver==1.3.2
MarkupPy==1.14
numpy==1.21.4
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==8.4.0
pycparser==2.20
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
setuptools-scm==6.3.2
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tomli==1.2.2
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
